backend/auth_app/adapters.py
import re
from allauth.socialaccount.adapter import DefaultSocialAccountAdapter


class CustomSocialAccountAdapter(DefaultSocialAccountAdapter):
    def is_auto_signup_allowed(self, request, sociallogin):
        return True

    def populate_user(self, request, sociallogin, data):
        user = super().populate_user(request, sociallogin, data)
        if not user.username:
            email = data.get('email') or ''
            base = re.sub(r'[^\w]', '_', email.split('@')[0])[:30] or 'user'
            from django.contrib.auth import get_user_model
            User = get_user_model()
            username = base
            counter = 1
            while User.objects.filter(username=username).exists():
                username = f"{base}_{counter}"
                counter += 1
            user.username = username
        return user
