backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
python-dotenv>=1.0
mysqlclient>=2.2
PyMuPDF>=1.24
pdfplumber>=0.11
pdf2image>=1.17
pytesseract>=0.3
python-docx>=1.1
Pillow>=10.3
requests>=2.31

# Google OAuth
google-auth>=2.29
google-auth-oauthlib>=1.2
google-auth-httplib2>=0.2

# django-allauth (Google OAuth via allauth)
django-allauth>=0.63
django-allauth[socialaccount]

# Async task queue
celery>=5.4
redis>=5.0
django-celery-results>=2.5
django-celery-beat>=2.6

# AI — Gemini
google-genai>=1.0

# Cloud storage
django-storages>=1.14
boto3>=1.34

# Security
django-axes>=8.0
django-ratelimit>=4.0
cryptography>=42.0
